Update and rename README.md to Code
The People who worked on this project were Josiah, Elijah, Ian, and Raven, From Chester County High School.
I, Josiah Cook did the coding, this file's description will include what I've done
I had a couple times in which I misspelled the code I was basing it off of, or times where the code seemed to be outdated and had to find a new alternative. Which is why I use dict.git in the code.
I Did not have to use a 3D printer for this project.
I enjoy Coding Class Very Much.